Redis, Security

+++ src/main/java/ru/morozov/users/controller/UsersController.java
@@ -4,33 +4,33 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.util.StringUtils;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 import ru.morozov.users.dto.NewUserDto;
 import ru.morozov.users.dto.UserDto;
 import ru.morozov.users.exceptions.NotFoundException;
 import ru.morozov.users.service.UserService;
+import ru.morozov.users.utils.AuthUtils;
 
 @RestController
-@RequestMapping
+@RequestMapping("/user")
 @RequiredArgsConstructor
 @Slf4j
 public class UsersController {
 
     private final UserService userService;
 
-    @PostMapping
-    @ResponseStatus(HttpStatus.CREATED)
-    public UserDto createUser(@RequestBody NewUserDto user) {
-        if (StringUtils.isEmpty(user.getUsername())) {
-            throw new RuntimeException("Empty username");
-        }
-
-        return userService.create(user);
+    @GetMapping("/me")
+    public UserDto getMe() {
+        return (UserDto) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
     }
 
     @GetMapping("/{userId:\\d+}")
     public ResponseEntity getUser(@PathVariable("userId") Long userId) {
+        if (!AuthUtils.getCurrentUserId().equals(userId)) {
+            return new ResponseEntity(HttpStatus.FORBIDDEN);
+        }
+
         try {
             return new ResponseEntity(
                     userService.get(userId),
@@ -44,6 +44,10 @@ public class UsersController {
 
     @DeleteMapping("/{userId:\\d+}")
     public ResponseEntity deleteUser(@PathVariable("userId") Long userId) {
+        if (!AuthUtils.getCurrentUserId().equals(userId)) {
+            return new ResponseEntity(HttpStatus.FORBIDDEN);
+        }
+
         try {
             userService.delete(userId);
             return new ResponseEntity(HttpStatus.OK);
@@ -55,6 +59,10 @@ public class UsersController {
 
     @PutMapping("/{userId:\\d+}")
     public ResponseEntity updateUser(@PathVariable("userId") Long userId, @RequestBody NewUserDto user) {
+        if (!AuthUtils.getCurrentUserId().equals(userId)) {
+            return new ResponseEntity(HttpStatus.FORBIDDEN);
+        }
+
         try {
             userService.update(userId, user);
             return new ResponseEntity(HttpStatus.OK);
